Use single parameters, make plotting test faster
Use single base type parameters instead of lists where possible in order
to make it easier to pass these from the command line, as well as use
the operations in workflows, without hackery.

Improved the test speed of the plotting test. Now there are two tests,
that together take around a second, if it's still to slow, it's possible
to switch the nominal test off and only test the exception throwing.