Remove file parameter
According to operation specification, the correlation operation should
take in a 'file' parameter if the user wants to save the results.

However, this unnecessarily complicates the operation signature which
translates to large and hard to understand GUI forms for operations.
Moreover, operations shouldn't really deal with io, except for some very
special cases.

This case is not such, as the operation returns either an xarray Dataset
 or (soon will return also a) pandas Dataframe, which can then be saved
as netCdf, csv or text files using dedicated io operations.