Remove 'type' parameter
In the first operation specification it was said that the correlation
operation should perform a 'pixel by pixel' or 'time by time'
correlation on 3D datasets.

'time by time' correlation of a 3D dataset would imply aggregating the
dataset beforehand. For example, taking a mean. As there can be many
ways how two datasets can be spatially aggregated, this should not
happen as part of a correlation operation, as this would unneccessarily
inflate the amount of parameters. Rather, such aggregations should be
done before hand, and two 1D timeseries passed to the correlation
operation instead.