Coerce chunking along time dimension
The ``open_dataset`` implementation only provides chunks along spatial
dimensions. This results in all other dimensions including 'time' not
getting chunked. This creates problems later on with retrieving the
computation results for viewing.

The 'time' dimension now gets chunked such that each time
slice is a different chunk, which is a good assumption in
most cases.

Closes #698